@@ -97,28 +97,28 @@
     "@angular/cli": "^19.2.9",
     "@angular/compiler-cli": "^19.2.8",
     "@angular/language-service": "^19.2.8",
-    "@capacitor/cli": "^7.2.0",
-    "@tailwindcss/postcss": "^4.1.4",
+    "@capacitor/cli": "^7.4.3",
+    "@tailwindcss/postcss": "^4.1.13",
     "@trapezedev/project": "^7.1.3",
     "@types/angular-gettext": "^2.1.38",
     "@types/d3": "^7.4.3",
     "@types/file-saver": "^2.0.7",
-    "@types/jasmine": "^5.1.4",
+    "@types/jasmine": "^5.1.9",
     "@types/lodash-es": "^4.17.12",
     "@types/mapbox__polyline": "^1.0.5",
-    "@types/node": "^22.15.2",
-    "@types/pako": "^2.0.3",
+    "@types/node": "^24.5.2",
+    "@types/pako": "^2.0.4",
     "@types/platform": "^1.3.6",
     "@types/proj4": "^2.5.5",
     "@types/uuid": "^10.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.31.0",
-    "@typescript-eslint/parser": "^8.31.0",
-    "eslint": "^9.25.1",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jsdoc": "^50.6.11",
+    "@typescript-eslint/eslint-plugin": "^8.44.1",
+    "@typescript-eslint/parser": "^8.44.1",
+    "eslint": "^9.36.0",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-jsdoc": "^60.5.0",
     "eslint-plugin-prefer-arrow": "^1.2.3",
-    "geojson-vt": "^3.2.1",
-    "jasmine-core": "^5.7.0",
+    "geojson-vt": "^4.0.2",
+    "jasmine-core": "^5.11.0",
     "jasmine-spec-reporter": "^7.0.0",
     "karma": "^6.4.4",
     "karma-chrome-launcher": "^3.2.0",
