@@ -1,7 +1,6 @@
 import { Injectable, EventEmitter, NgZone, inject } from "@angular/core";
-import { MapMouseEvent, Map, GeoJSONSource } from "maplibre-gl";
+import { MapMouseEvent, Map, GeoJSONSource, type Point } from "maplibre-gl";
 import { Store } from "@ngxs/store";
-import type Point from "@mapbox/point-geometry";
 import type { Immutable } from "immer";
 
 import { SelectedRouteService } from "../../services/selected-route.service";
