@@ -1,5 +1,5 @@
 ﻿import { Urls } from "../urls";
-import { environment } from "environments/environment";
+import { environment } from "../../environments/environment";
 import type { Language, MutableApplicationState, RouteData, StateWithHistory } from "../models";
 
 export const ISRAEL_HIKING_MAP = "Israel Hiking Map";
